+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/FeatureGates.java
@@ -14,9 +14,14 @@ import static java.util.Arrays.asList;
  * Class for handling the configuration of feature gates
  */
 public class FeatureGates {
-    private static final String CONTROL_PLANE_LISTENER = "ControlPlaneListener";
+    public static final FeatureGates NONE = new FeatureGates("");
 
+    private static final String CONTROL_PLANE_LISTENER = "ControlPlaneListener";
+    private static final String SERVICE_ACCOUNT_PATCHING = "ServiceAccountPatching";
+
+    // When adding new feature gates, do not forget to add them to allFeatureGates() and toString() methods
     private final FeatureGate controlPlaneListener = new FeatureGate(CONTROL_PLANE_LISTENER, false);
+    private final FeatureGate serviceAccountPatching = new FeatureGate(SERVICE_ACCOUNT_PATCHING, false);
 
     /**
      * Constructs the feature gates configuration.
@@ -40,6 +45,9 @@ public class FeatureGates {
                 switch (featureGate) {
                     case CONTROL_PLANE_LISTENER:
                         setValueOnlyOnce(controlPlaneListener, value);
+                        break;
+                    case SERVICE_ACCOUNT_PATCHING:
+                        setValueOnlyOnce(serviceAccountPatching, value);
                         break;
                     default:
                         throw new InvalidConfigurationException("Unknown feature gate " + featureGate + " found in the configuration");
@@ -70,10 +78,30 @@ public class FeatureGates {
         return controlPlaneListener.isEnabled();
     }
 
+    /**
+     * @return  Returns true when the ServiceAccountPatching feature gate is enabled
+     */
+    public boolean serviceAccountPatchingEnabled() {
+        return serviceAccountPatching.isEnabled();
+    }
+
+    /**
+     * Returns a list of all Feature gates. Used for testing.
+     *
+     * @return  List of all Feature Gates
+     */
+    /*test*/ List<FeatureGate> allFeatureGates()  {
+        return List.of(
+                controlPlaneListener,
+                serviceAccountPatching
+        );
+    }
+
     @Override
     public String toString() {
         return "FeatureGates(" +
-                "controlPlaneListener=" + controlPlaneListener.isEnabled() +
+                "controlPlaneListener=" + controlPlaneListener.isEnabled() + "," +
+                "ServiceAccountPatching=" + serviceAccountPatching.isEnabled() +
                 ")";
     }
 
@@ -124,6 +152,13 @@ public class FeatureGates {
          */
         public boolean isEnabled() {
             return value == null ? defaultValue : value;
+        }
+
+        /**
+         * @return  Returns True if this feature gate is enabled by default. False otherwise.
+         */
+        public boolean isEnabledByDefault() {
+            return defaultValue;
         }
     }
 }