+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/operator/resource/ResourceOperatorSupplier.java
@@ -22,6 +22,7 @@ import io.strimzi.api.kafka.model.KafkaMirrorMaker;
 import io.strimzi.api.kafka.model.KafkaMirrorMaker2;
 import io.strimzi.api.kafka.model.KafkaRebalance;
 import io.strimzi.operator.PlatformFeaturesAvailability;
+import io.strimzi.operator.cluster.FeatureGates;
 import io.strimzi.operator.common.AdminClientProvider;
 import io.strimzi.operator.common.BackOff;
 import io.strimzi.operator.common.DefaultAdminClientProvider;
@@ -91,7 +92,7 @@ public class ResourceOperatorSupplier {
     public final MetricsProvider metricsProvider;
     public AdminClientProvider adminClientProvider;
 
-    public ResourceOperatorSupplier(Vertx vertx, KubernetesClient client, PlatformFeaturesAvailability pfa, long operationTimeoutMs) {
+    public ResourceOperatorSupplier(Vertx vertx, KubernetesClient client, PlatformFeaturesAvailability pfa, FeatureGates gates, long operationTimeoutMs) {
         this(vertx, client,
             new ZookeeperLeaderFinder(vertx, new SecretOperator(vertx, client),
             // Retry up to 3 times (4 attempts), with overall max delay of 35000ms
@@ -99,12 +100,12 @@ public class ResourceOperatorSupplier {
                     new DefaultAdminClientProvider(),
                     new DefaultZookeeperScalerProvider(),
                     new MicrometerMetricsProvider(),
-                    pfa, operationTimeoutMs);
+                    pfa, gates, operationTimeoutMs);
     }
 
     public ResourceOperatorSupplier(Vertx vertx, KubernetesClient client, ZookeeperLeaderFinder zlf,
                                     AdminClientProvider adminClientProvider, ZookeeperScalerProvider zkScalerProvider,
-                                    MetricsProvider metricsProvider, PlatformFeaturesAvailability pfa, long operationTimeoutMs) {
+                                    MetricsProvider metricsProvider, PlatformFeaturesAvailability pfa, FeatureGates gates, long operationTimeoutMs) {
         this(new ServiceOperator(vertx, client),
                 pfa.hasRoutes() ? new RouteOperator(vertx, client.adapt(OpenShiftClient.class)) : null,
                 new ZookeeperSetOperator(vertx, client, zlf, operationTimeoutMs),
@@ -113,7 +114,7 @@ public class ResourceOperatorSupplier {
                 new SecretOperator(vertx, client),
                 new PvcOperator(vertx, client),
                 new DeploymentOperator(vertx, client),
-                new ServiceAccountOperator(vertx, client),
+                new ServiceAccountOperator(vertx, client, gates.serviceAccountPatchingEnabled()),
                 new RoleBindingOperator(vertx, client),
                 new RoleOperator(vertx, client),
                 new ClusterRoleBindingOperator(vertx, client),