+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/model/KafkaConnectClusterTest.java
@@ -26,6 +26,7 @@ import io.fabric8.kubernetes.api.model.SecretVolumeSourceBuilder;
 import io.fabric8.kubernetes.api.model.SecurityContext;
 import io.fabric8.kubernetes.api.model.SecurityContextBuilder;
 import io.fabric8.kubernetes.api.model.Service;
+import io.fabric8.kubernetes.api.model.ServiceAccount;
 import io.fabric8.kubernetes.api.model.TopologySpreadConstraint;
 import io.fabric8.kubernetes.api.model.TopologySpreadConstraintBuilder;
 import io.fabric8.kubernetes.api.model.Volume;
@@ -543,6 +544,9 @@ public class KafkaConnectClusterTest {
         Map<String, String> crbLabels = TestUtils.map("l9", "v9", "l10", "v10");
         Map<String, String> crbAnots = TestUtils.map("a9", "v9", "a10", "v10");
 
+        Map<String, String> saLabels = TestUtils.map("l11", "v11", "l12", "v12");
+        Map<String, String> saAnots = TestUtils.map("a11", "v11", "a12", "v12");
+
         HostAlias hostAlias1 = new HostAliasBuilder()
                 .withHostnames("my-host-1", "my-host-2")
                 .withIp("192.168.1.86")
@@ -608,6 +612,12 @@ public class KafkaConnectClusterTest {
                                 .withAnnotations(crbAnots)
                             .endMetadata()
                         .endClusterRoleBinding()
+                        .withNewServiceAccount()
+                            .withNewMetadata()
+                                .withLabels(saLabels)
+                                .withAnnotations(saAnots)
+                            .endMetadata()
+                        .endServiceAccount()
                     .endTemplate()
                 .endSpec()
                 .build();
@@ -645,6 +655,11 @@ public class KafkaConnectClusterTest {
         ClusterRoleBinding crb = kc.generateClusterRoleBinding();
         assertThat(crb.getMetadata().getLabels().entrySet().containsAll(crbLabels.entrySet()), is(true));
         assertThat(crb.getMetadata().getAnnotations().entrySet().containsAll(crbAnots.entrySet()), is(true));
+
+        // Check Service Account
+        ServiceAccount sa = kc.generateServiceAccount();
+        assertThat(sa.getMetadata().getLabels().entrySet().containsAll(saLabels.entrySet()), is(true));
+        assertThat(sa.getMetadata().getAnnotations().entrySet().containsAll(saAnots.entrySet()), is(true));
     }
 
     @ParallelTest