+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## 0.24.0
 
 * Add support for Kafka 2.7.1 and remove support for 2.6.0, 2.6.1, and 2.6.2
+* Support for patching of service accounts and configuring their labels and annotations. The feature is disabled by default and enabled using the new `ServiceAccountPatching` feature gate.
 * Added support for configuring cluster-operator's worker thread pool size that is used for various sync and async tasks
- 
+
 ## 0.23.0
 
 * Add support for Kafka 2.8.0 and 2.6.2, remove support for Kafka 2.5.x